% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_salary.R
\docType{data}
\name{average_salary}
\alias{average_salary}
\title{Salary Data for a Hypothetical Firm}
\format{A data frame with 25 rows and 1 columns.

\tabular{lll}{
name \tab type \tab description \cr
\code{salary} \tab \code{"numeric"} \tab Salary in thousands of dollars \cr
}}
\source{
Hoff, D. (1954) \emph{How to Lie with Statistics}. W.W. Norton & Company.
}
\usage{
average_salary
}
\description{
Salary data for a hypothetical firm used in Darrell Hoff's
\emph{How to Lie with Statistics} to illustrate the difference
between the mean, median, and mode.
}
\keyword{datasets}
