% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countydeaths.R
\docType{data}
\name{countydeaths}
\alias{countydeaths}
\title{County death rates for six English counties, 1876-1883}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 48 rows and 3 columns.}
\source{
Stigler, Stephen M. \emph{The Seven Pillars of Statistical Wisdom}, p. 98
}
\usage{
countydeaths
}
\description{
County deaths for six English counties, 1876--1883, from the 1883 registrar-general's report. This data was analyzed in Edgeworth (1885).
}
\details{
A data frame with 3 columns and 48 rows:
\describe{
 \item{ \code{ county }}{ \code{ character }. County name }
 \item{ \code{ year }}{ \code{ integer }. Year }
 \item{ \code{ death_rate }}{ \code{ numeric }. Annual deaths per thousand in that county }

}

A data frame with 3 columns and 48 rows:
\describe{
 \item{ \code{ county }}{ \code{ character }. County name }
 \item{ \code{ year }}{ \code{ integer }. Year }
 \item{ \code{ death_rate }}{ \code{ numeric }. Annual deaths per thousand in that county }

}

A data frame with 3 columns and 48 rows:
\describe{
 \item{ \code{ county }}{ \code{ character }. County name }
 \item{ \code{ year }}{ \code{ integer }. Year }
 \item{ \code{ death_rate }}{ \code{ numeric }. Annual deaths per thousand in that county }

}
}
\references{
\itemize{
\item{ Stigler, Stephen M. \emph{The Seven Pillars of Statistical Wisdom}, p. 26 }
\item{ Edgeworth, Francis Ysidro. 1885. \dQuote{On Methods of Ascertaining Variations in the Rate of Births, Deaths and Marriages}. \emph{Journal of the [Royal] Statistical Society}. 48: 628-649 }

}

\itemize{
\item{ Stigler, Stephen M. \emph{The Seven Pillars of Statistical Wisdom}, p. 26 }
\item{ Edgeworth, Francis Ysidro. 1885. \dQuote{On Methods of Ascertaining Variations in the Rate of Births, Deaths and Marriages}. \emph{Journal of the [Royal] Statistical Society}. 48: 628-649 }

}
}
\keyword{datasets}
