% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delhiwheat.R
\docType{data}
\name{delhiwheat}
\alias{delhiwheat}
\title{Price of Wheat in Delhi 1763-1836}
\format{A data frame with 74 rows and 4 columns.

\tabular{lll}{
name \tab type \tab description \cr
\code{year} \tab \code{"integer"} \tab Year \cr
\code{price} \tab \code{"integer"} \tab Price in rupees per 1,000 seers of wheat in Delhi \cr
\code{maximum} \tab \code{"logical"} \tab Indicator marking a price maxima in a cycle \cr
\code{crisis} \tab \code{"logical"} \tab Commercial crisis \cr
}}
\source{
Jevons, W. Stanley. 1879. "Sun-Spots and Commercial Crises." \emph{Nature}. \href{https://dx.doi.org/10.1038/019588a0}{doi:10.1038/019588a0}.
}
\usage{
delhiwheat
}
\description{
Price of wheat in Delhi 1763-1836.  This series was notoriously and spuriously related to sunspots (Wolf's numbers) by William S. Jevons in his last published paper. Jevons theorized that sunspots were related to commercial cycles and crises due to similarity in cycles of these time series.
}
\references{
Stigler, Stephen M. \emph{The Seven Pillars of Statistical Wisdom}

Jevons, W. Stanley. 1882. "The Solar-Commercial Cycle." \emph{Nature}. \href{https://dx.doi.org/10.1038/026226a0}{doi:10.1038/026226a0}.
}
\seealso{
\code{nonsense} for another example of spurious correlations.
}
\keyword{datasets}
