% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eponymns.R
\docType{data}
\name{eponymns}
\alias{eponymns}
\title{Data on}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 80 rows and 5 columns.}
\usage{
eponymns
}
\description{
Data from eighty textbooks, 1816 to 1976, on their eponymic practices regarding the normal distribution. Each text was coded as whether they called the normal distribution "Gaussian", "Laplacian", or did not use an eponymn.
}
\details{
A data frame with 5 columns and 80 rows:
\describe{
 \item{ \code{ year }}{ \code{ integer }. Year the text was published. }
 \item{ \code{ country }}{ \code{ character }. Country in which the text was published. }
 \item{ \code{ gaussian }}{ \code{ logical }. Did the text refer to the normal distribution as "Gaussian"? }
 \item{ \code{ laplace }}{ \code{ logical }. Did the text refer to the normal distribution as "Laplace"? }
 \item{ \code{ noneponymic }}{ \code{ logical }. Did the text not use an eponymn when referring to the normal distribution? }

}

A data frame with 5 columns and 80 rows:
\describe{
 \item{ \code{ year }}{ \code{ integer }. Year the text was published. }
 \item{ \code{ country }}{ \code{ character }. Country in which the text was published. }
 \item{ \code{ gaussian }}{ \code{ logical }. Did the text refer to the normal distribution as "Gaussian"? }
 \item{ \code{ laplace }}{ \code{ logical }. Did the text refer to the normal distribution as "Laplace"? }
 \item{ \code{ noneponymic }}{ \code{ logical }. Did the text not use an eponymn when referring to the normal distribution? }

}

A data frame with 5 columns and 80 rows:
\describe{
 \item{ \code{ year }}{ \code{ integer }. Year the text was published. }
 \item{ \code{ country }}{ \code{ character }. Country in which the text was published. }
 \item{ \code{ gaussian }}{ \code{ logical }. Did the text refer to the normal distribution as "Gaussian"? }
 \item{ \code{ laplace }}{ \code{ logical }. Did the text refer to the normal distribution as "Laplace"? }
 \item{ \code{ noneponymic }}{ \code{ logical }. Did the text not use an eponymn when referring to the normal distribution? }

}

A data frame with 5 columns and 80 rows:
\describe{
 \item{ \code{ year }}{ \code{ integer }. Year the text was published. }
 \item{ \code{ country }}{ \code{ character }. Country in which the text was published. }
 \item{ \code{ gaussian }}{ \code{ logical }. Did the text refer to the normal distribution as "Gaussian"? }
 \item{ \code{ laplace }}{ \code{ logical }. Did the text refer to the normal distribution as "Laplace"? }
 \item{ \code{ noneponymic }}{ \code{ logical }. Did the text not use an eponymn when referring to the normal distribution? }

}

A data frame with 5 columns and 80 rows:
\describe{
 \item{ \code{ year }}{ \code{ integer }. Year the text was published. }
 \item{ \code{ country }}{ \code{ character }. Country in which the text was published. }
 \item{ \code{ gaussian }}{ \code{ logical }. Did the text refer to the normal distribution as "Gaussian"? }
 \item{ \code{ laplace }}{ \code{ logical }. Did the text refer to the normal distribution as "Laplace"? }
 \item{ \code{ noneponymic }}{ \code{ logical }. Did the text not use an eponymn when referring to the normal distribution? }

}
}
\keyword{datasets}
