% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kerrich.R
\docType{data}
\name{kerrich}
\alias{kerrich}
\title{John Kerrich's Coin Tossing Experiment (Summary)}
\format{A data frame with 28 rows and 2 columns.

\tabular{lll}{
name \tab type \tab description \cr
\code{tosses} \tab \code{"integer"} \tab number of tosses \cr
\code{heads} \tab \code{"integer"} \tab Number of heads \cr
}}
\usage{
kerrich
}
\description{
John Kerrich, along with internee Eric Christensen, demonstrated the Law of Large Numbers by tossing a coin 10,000 times and recording the number of heads while interned in Denmark during World War II. In total, the tosses yielded 5,067 heads. Kerrich discussed these results in the text \emph{An experimental introduction to the theory of probability}.
This data provides a summary of the coin tosses with the number of heads heads at certain intervals of number of tosses: 10-90 by 10, 100-900 by 100, and 1000-10,000 by 1,000.
}
\references{
\url{https://stats.stackexchange.com/questions/76663/john-kerrich-coin-flip-data}

\url{https://en.wikipedia.org/wiki/John_Edmund_Kerrich#cite_note-6}

\url{https://web.archive.org/web/20140715063158/http://www.wits.ac.za/academic/science/stats/school/5717/history.html}

Kerich, J. E. (1946) \emph{An Experimental Introduction to the Theory of Probability}. Copenhagen:E. Munksgaard.
}
\keyword{datasets}
