% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_deviates.R
\docType{data}
\name{normal_deviates}
\alias{normal_deviates}
\title{100,000 Random Normal Deviates}
\format{An object of class \code{numeric} of length 100000.}
\source{
\url{https://www.rand.org/content/dam/rand/pubs/monograph_reports/MR1418/MR1418.deviates.txt.zip}
}
\usage{
normal_deviates
}
\description{
100,000 random normal deviates generated by the RAND Corporation in 1955 in the book \emph{A Million Random Digits with 100,000 Normal Deviates}.
}
\references{
RAND (1955) \emph{A Million Random Digits with 100,000 Normal Deviates}. \url{https://www.rand.org/pubs/monograph_reports/MR1418/index2.html}
}
\keyword{datasets}
