% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlanetMoneyCowGuesses.R
\docType{data}
\name{PlanetMoneyCowGuesses}
\alias{PlanetMoneyCowGuesses}
\title{mph{Planet Money} Cow Weight Guesses}
\format{A data frame with 2 columns and 17184 rows:
\describe{
 \item{ \code{ guess }}{ \code{ numeric }. Response to the question, "How much does this cow weigh?". The weights are in pounds. } \item{ \code{ error }}{ \code{ numeric }. Difference (in pounds) between the guess and the weight of 1,355 pounds. }
}}
\source{
\href{https://docs.google.com/spreadsheets/d/1tLZpxu0FzX5v6Dq1BYM4RWgdFygit4En6L8sldMz68Q/edit?usp=sharing}{Planet Money Cow Data}
}
\usage{
PlanetMoneyCowGuesses
}
\description{
Data collected by the NPR podcast \emph{Planet Money} to the question, "How much does this cow weigh?". This is a replication of the Galton paper "Vox Populi" which cites guesses of cow's weight from a county fair to illustrate the "wisdom of crowds".
}
\references{
\itemize{
\item{ Quoctrung Bui, &quot;17,205 People Guessed The Weight Of A Cow. Here's How They Did.&quot;, \emph{Planet Money}, August 7, 2015. \url{http://www.npr.org/sections/money/2015/08/07/429720443/17-205-people-guessed-the-weight-of-a-cow-heres-how-they-did}. }
\item{ Jacob Goldstein, &quot;How Much Does This Cow Weigh?,&quot; \emph{Planet Money}, July 17, 2015. }

}

\itemize{
\item{ Quoctrung Bui, &quot;17,205 People Guessed The Weight Of A Cow. Here's How They Did.&quot;, \emph{Planet Money}, August 7, 2015. \url{http://www.npr.org/sections/money/2015/08/07/429720443/17-205-people-guessed-the-weight-of-a-cow-heres-how-they-did}. }
\item{ Jacob Goldstein, &quot;How Much Does This Cow Weigh?,&quot; \emph{Planet Money}, July 17, 2015. }

}
}
\keyword{datasets}
