% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_digits.R
\docType{data}
\name{rand_digits}
\alias{rand_digits}
\title{One Million Random Digits}
\format{An object of class \code{integer} of length 1000000.}
\source{
\url{https://www.rand.org/content/dam/rand/pubs/monograph_reports/MR1418/MR1418.digits.txt.zip}
}
\usage{
rand_digits
}
\description{
One million random digits generated by the RAND Corporation in 1955 in the book \emph{A Million Random Digits with 100,000 Normal Deviates}.
}
\references{
RAND (1955) \emph{A Million Random Digits with 100,000 Normal Deviates}. \url{https://www.rand.org/pubs/monograph_reports/MR1418/index2.html}
}
\keyword{datasets}
