% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VoxPopuli.R
\docType{data}
\name{VoxPopuli}
\alias{VoxPopuli}
\title{Galton's cow weight-judging competition data}
\format{A data frame with 19 rows and 3 columns.

\tabular{lll}{
name \tab type \tab description \cr
\code{quantile} \tab \code{"numeric"} \tab Quantile of the distribution of guesses. \cr
\code{estimate} \tab \code{"numeric"} \tab Value of the guess (lbs.) \cr
\code{normal} \tab \code{"numeric"} \tab Quantile values of a normal distribution with a median of 1,207 lbs. and a probable error (half the interquartile range) of 37 lbs (standard deviation of 25 lbs).
 \cr
}}
\source{
Galton, Francis. 1907. "Vox Populi." \emph{Nature}. \href{http://dx.doi.org/10.1038/075414a0}{doi:10.1038/075414a0}
}
\usage{
VoxPopuli
}
\description{
Distribution of 787 guesses in a cow weight-judging competition.
These data was used by Galton to in a series of letters to
\emph{Nature} to illustrate the wisdom of crowds.
}
\details{
The actual weight of the cow was 1,198 pounds.
}
\references{
Galton, Francis. 1907. "One Value, One Vote." \emph{Nature}. \href{http://dx.doi.org/10.1038/075414a0}{doi:10.1038/075414a0}

Galton, Francis. 1907. "Ballot Box." \emph{Nature} \href{http://dx.doi.org/10.1038/075509e0}{doi:10.1038/075509e0}

Hooke, R. H. 1907. "Mean or Median." \emph{Nature}.  \href{http://dx.doi.org/10.1038/075487a0}{doi:10.1038/075487a0}

Yule, G. Udny. 1907. "Mean or Median." \emph{Nature}.  \href{http://dx.doi.org/10.1038/075534b0}{doi:10.1038/075534b0}

Kenneth W. Wallis. 2014. "Revisiting Francis Galton's Forecasting Competition" \url{https://arxiv.org/pdf/1410.3989.pdf}

Levy, D. M. and Peart, S. (2002). Galton's two papers on voting as robust estimation. \emph{Public Choice} 113 357–365.
}
\keyword{datasets}
