% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winetasting.R
\docType{data}
\name{winetasting}
\alias{winetasting}
\title{Paris Wine Tasting}
\format{An object of class \code{list} of length 3.}
\source{
\url{http://www.liquidasset.com/lindley.htm}
}
\usage{
winetasting
}
\description{
Judges scores from the Paris Wine Tasting or 1976. In this taste testing, French judges conducted blind taste testings of Chardonnay and Cabernet Sauvignon wines from France and California. Surprisingly, California wines won each category.
A named list of three data frames: \itemize{ \item{\code{scores}}{The judges' scores for each test.} \item{\code{judges}}{The names of the judges.} \item{\code{wines}}{The wines that were tasted.} }
}
\keyword{datasets}
