% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bordeaux.R
\docType{data}
\name{bordeaux}
\alias{bordeaux}
\title{Prices of Bordeaux Vintages}
\format{A data frame with 38 rows and 6 columns.

\tabular{lll}{
name \tab type \tab description \cr
\code{vint} \tab \code{"integer"} \tab Vintage year \cr
\code{price} \tab \code{"numeric"} \tab Price of the vintage, as an index (1961=100). \cr
\code{wrain} \tab \code{"integer"} \tab Winter (October to March) rain in ML. \cr
\code{degrees} \tab \code{"numeric"} \tab Average temperature (degrees centigrade) April to September. \cr
\code{hrain} \tab \code{"integer"} \tab Harvest (August and September) rain in ML \cr
\code{time_sv} \tab \code{"integer"} \tab Time since vintage - relative to 1983 \cr
}}
\source{
\url{http://www.liquidasset.com/winedata.html}
}
\usage{
bordeaux
}
\description{
Prices of Vintages of Bordeaux Wines in New York auctions. The price is an index based on the wines of multiple chateaus, from a regression of the logarithm price from several thousand auction sales with variables for chateau and vintage years. These data include vintages 1952 to 1980, with missing price values for 1954 and 1956 because they were rarely sold.
}
\references{
Ashenfelter, O. and Ashmore, D. and Lalonde, R. (1995) "Bordeaux Wine Vintage Quality and the Weather." \emph{Chance}. \href{https://dx.doi.org/10.1080/09332480.1995.10542468}{doi:10.1080/09332480.1995.10542468}

Ashenfelter, O. (2008). "Predicting the Quality and Prices of Bordeaux Wine." \emph{The Economic Journal} \href{https://dx.doi.org/10.1111/j.1468-0297.2008.02148.x}{doi:10.1111/j.1468-0297.2008.02148.x}
}
\keyword{datasets}
