% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borough.R
\docType{data}
\name{borough}
\alias{borough}
\title{Borough's 1580 data for the variation of the needle}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9 rows and 4 columns.}
\source{
Stigler, Stephen M. \emph{The Seven Pillars of Statistical Wisdom}, p. 20
}
\usage{
borough
}
\description{
William Borough's data for readings of the variation between magnetic and
}
\details{
true north using the Sun's taken at different times of the day at

the Limehouse lighthouse, not far from the Greenwich Meridian.

These readings was published in \emph{A Discours of the Variation of the Campas, or Maneticall Needle}.



A data frame with 4 columns and 9 rows:
\describe{
 \item{ \code{ elevation }}{ \code{ integer }. Elevation of the Sun in degrees }
 \item{ \code{ variation_am }}{ \code{ numeric }. Difference between the shadow of the Sun before noon and the compass north. }
 \item{ \code{ variation_pm }}{ \code{ numeric }. Difference between the shadow of the Sun after noon and the compass north. }
 \item{ \code{ variation_north }}{ \code{ numeric }. Variation between true and magnetic north. }

}
}
\references{
\itemize{
\item{ Stigler, Stephen M. \emph{The Seven Pillars of Statistical Wisdom} }

}
}
\keyword{datasets}
