% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db1668.R
\docType{data}
\name{db1668}
\alias{db1668}
\title{Table of Magnetic Variation near Bristol on May 23, 1668}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 4 columns.}
\source{
Stigler, Stephen M. \emph{The Seven Pillars of Statistical Wisdom}
}
\usage{
db1668
}
\description{
This is from a note published in the \emph{Transactions of the Royal Society} in 1668. The editor, Henry Oldenburg,  printed parts of a letter from "D.B." with five values of the mangetic variation near Bristol, England on May 23, 1668. This letter is notable in that he calculates the "mean of the true Variation", where "mean" refers to the artithmetic mean. It is referenced by Stigler as the first instance in which the term "mean" is used to refer to the arithmetical mean. D.B. reported a mean of 1 degree 27 minutes, rounding down from the arithmetic mean of 1 degree 28 minutes.
}
\details{
A data frame with 4 columns and 5 rows:
\describe{
 \item{ \code{ sun_altitude }}{ \code{ numeric }. Sun's altitude in degrees }
 \item{ \code{ magnetic_azimuth }}{ \code{ numeric }. Magnetic azimuth in degrees }
 \item{ \code{ sun_azimuth }}{ \code{ numeric }. Sun's azimuth in degrees }
 \item{ \code{ variation }}{ \code{ numeric }. Variation in degrees westerly, between true north (sun azimuth) and magnetic north (magnetic azimuth) }

}
}
\references{
\itemize{
\item{ Stigler, Stephen M. \emph{The Seven Pillars of Statistical Wisdom}, p. 26 }

}
}
\keyword{datasets}
