% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delhiwheat.R
\docType{data}
\name{delhiwheat}
\alias{delhiwheat}
\title{Price of Wheat in Delhi 1763-1836}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 74 rows and 4 columns.}
\source{
Jevons, W. Stanley. 1879. "Sun-Spots and Commercial Crises." \emph{Nature}. \href{https://dx.doi.org/10.1038/019588a0}{doi:10.1038/019588a0}.
}
\usage{
delhiwheat
}
\description{
Price of wheat in Delhi 1763-1836.  This series was notoriously and spuriously related to sunspots (Wolf's numbers) by William S. Jevons in his last published paper. Jevons theorized that sunspots were related to commercial cycles and crises due to similarity in cycles of these time series.
}
\details{
A data frame with 4 columns and 74 rows:
\describe{
 \item{ \code{ year }}{ \code{ integer }. Year }
 \item{ \code{ price }}{ \code{ integer }. Price in rupees per 1,000 seers of wheat in Delhi }
 \item{ \code{ maximum }}{ \code{ logical }. Indicator marking a price maxima in a cycle }
 \item{ \code{ crisis }}{ \code{ logical }. Commercial crisis }

}

A data frame with 4 columns and 74 rows:
\describe{
 \item{ \code{ year }}{ \code{ integer }. Year }
 \item{ \code{ price }}{ \code{ integer }. Price in rupees per 1,000 seers of wheat in Delhi }
 \item{ \code{ maximum }}{ \code{ logical }. Indicator marking a price maxima in a cycle }
 \item{ \code{ crisis }}{ \code{ logical }. Commercial crisis }

}

A data frame with 4 columns and 74 rows:
\describe{
 \item{ \code{ year }}{ \code{ integer }. Year }
 \item{ \code{ price }}{ \code{ integer }. Price in rupees per 1,000 seers of wheat in Delhi }
 \item{ \code{ maximum }}{ \code{ logical }. Indicator marking a price maxima in a cycle }
 \item{ \code{ crisis }}{ \code{ logical }. Commercial crisis }

}

A data frame with 4 columns and 74 rows:
\describe{
 \item{ \code{ year }}{ \code{ integer }. Year }
 \item{ \code{ price }}{ \code{ integer }. Price in rupees per 1,000 seers of wheat in Delhi }
 \item{ \code{ maximum }}{ \code{ logical }. Indicator marking a price maxima in a cycle }
 \item{ \code{ crisis }}{ \code{ logical }. Commercial crisis }

}
}
\references{
\itemize{
\item{ Stigler, Stephen M. \emph{The Seven Pillars of Statistical Wisdom} }
\item{ Jevons, W. Stanley. 1882. &quot;The Solar-Commercial Cycle.&quot; \emph{Nature}. \href{https://dx.doi.org/10.1038/026226a0}{doi:10.1038/026226a0}. }

}

\itemize{
\item{ Stigler, Stephen M. \emph{The Seven Pillars of Statistical Wisdom} }
\item{ Jevons, W. Stanley. 1882. &quot;The Solar-Commercial Cycle.&quot; \emph{Nature}. \href{https://dx.doi.org/10.1038/026226a0}{doi:10.1038/026226a0}. }

}
}
\keyword{datasets}
