% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gavarett.R
\docType{data}
\name{gavarett}
\alias{gavarett}
\title{Gavarett's data on sex ratios of legimate and illegitimate births}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 3 columns.}
\source{
Stigler, Stephen M. \emph{The Seven Pillars of Statistical Wisdom}, p. 77-79
}
\usage{
gavarett
}
\description{
Jule's Gavarett data on the sex ratios of "legitimate" and illegitimate births for France 1824-1825. Gavarett is an early example of a significance test.
}
\details{
A data frame with 3 columns and 4 rows:
\describe{
 \item{ \code{ female }}{ \code{ integer }. 1 if female, 0 if male. }
 \item{ \code{ legitimate }}{ \code{ integer }. 1 if a "legitimate birth", 0 if a "illegitimate birth" }
 \item{ \code{ count }}{ \code{ integer }. Counts of births }

}
}
\references{
\itemize{
\item{ Stigler, Stephen M. \emph{The Seven Pillars of Statistical Wisdom} }
\item{ Gavarett, Jules. 1840. \emph{Pricipes generaux de statistique medicale}. Paris: Bechet jenue et Labe. }

}
}
\keyword{datasets}
