% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winetasting.R
\docType{data}
\name{winetasting}
\alias{winetasting}
\title{Paris Wine Tasting}
\format{An object of class \code{list} of length 3.}
\source{
\url{http://www.liquidasset.com/lindley.htm}
}
\usage{
winetasting
}
\description{
Judges scores from the Paris Wine Tasting or 1976. This consisted of blind taste testings of Chardonnay and Cabernet Sauvignons by French judges. In a surprise at the time, a Californian wine won each category, establishing California as a major wine growing region.
A named list of three data frames: \itemize{ \item{\code{scores}}{The judges' scores for each test.} \item{\code{judges}}{The names of the judges.} \item{\code{wines}}{The wines that were tasted.} }
}
\keyword{datasets}
